Tidy folder and run iconics notebook
